added some code examples of things I wrote for one-off problems or training purposes
numPrimesBeforeN - shows the number of prime numbers preceding a number
: this was first script I wrote a few years ago when learning about
creating custom classes and the initialize method being invoked when
running object_name.new()  montross_queue.rb - this was my solution to a
problem I was given to write a program that accepted queue instructions
and returned a reordered queue.  Detailed description in the file.
RubySelfExamles - this shows difference between class and instance
methods as well as provides an example of the prototype object in
javascript - how to modify the string object with the prototype object
by adding a new method or parameter.